backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
httpx==0.28.1
sqlalchemy==2.0.36
pymysql==1.1.1
python-dotenv==1.0.1
pydantic-settings==2.6.1
boto3>=1.35.0
